Change score to avg reads mapped on length

+++ src/build_bridge.h
@@ -72,7 +72,7 @@ void join_middle_edge(struct asm_edge_t e1, struct asm_edge_t e2,
 		struct asm_edge_t middle, char **res_seq);
 void get_path_scores(struct opt_proc_t *opt, struct asm_graph_t *g,
 		struct asm_graph_t *lg, struct path_info_t *pinfo,
-		int e1, int e2, int **scores);
+		int e1, int e2, float **scores);
 void join_bridge_center_by_path(struct asm_graph_t *lg, int *path, int path_len,
 		char **seq);
 void unrelated_filter(struct asm_edge_t e1, struct asm_edge_t e2,
